refactor: [030-search] 뉴스 검색 수정

+++ src/main/java/project/votebackend/dto/vote/VoteSearchResponse.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 public class VoteSearchResponse {
     private Long id;
     private String title;
-    private int totalVotes;
-    private int likeCount;
-    private int commentCount;
+    private long totalVotes;
+    private long likeCount;
+    private long commentCount;
 }
